refactor: move services. Rename folder
