+++ src/services/similarYear/similar-year.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { SimilarYearService } from './similar-year.service';
-import { OmdbApiModule } from '../shared/omdb-api/omdb-api.module';
+import { OmdbApiModule } from '../omdb-api/omdb-api.module';
 
 @Module({
   providers: [SimilarYearService],
